added boolean for making zip files

+++ test/wappsto_test.py
@@ -216,7 +216,7 @@ def validate_json(json_schema, arg):
         return False
 
 
-def set_up_log(self, log_file_exists, file_size):
+def set_up_log(self, log_file_exists, make_zip, file_size):
     """
     Sets up logs.
 
@@ -225,6 +225,7 @@ def set_up_log(self, log_file_exists, file_size):
     Args:
         self: referece to calling object
         log_file_exists: boolean indicating if log file should exist
+        make_zip: boolean indicating if log file should be zipped
         file_size: how big is the current size of the folder
 
     Returns:
@@ -247,9 +248,10 @@ def set_up_log(self, log_file_exists, file_size):
             string = ("0" * num_chars + "\n") * 1024
             file.write(string)
 
-        with zipfile.ZipFile(file_path.replace(".txt", ".zip"), "w") as zip_file:
-            zip_file.write(file_path, file_name)
-        os.remove(file_path)
+        if make_zip:
+            with zipfile.ZipFile(file_path.replace(".txt", ".zip"), "w") as zip_file:
+                zip_file.write(file_path, file_name)
+            os.remove(file_path)
 
     with open(self.service.event_storage.get_file_path("2000-1.txt"), "w") as file:
         file.write("")
@@ -390,9 +392,10 @@ class TestConnClass:
     @pytest.mark.parametrize("log_location", ["test_logs/logs"])
     @pytest.mark.parametrize("log_file_exists", [True, False])
     @pytest.mark.parametrize("load_from_state_file", [True, False])
+    @pytest.mark.parametrize("make_zip", [True, False])
     def test_connection(self, address, port, expected_status, callback_exists,
                         value_changed_to_none, upgradable, valid_json, log_offline,
-                        log_location, log_file_exists, load_from_state_file):
+                        log_location, log_file_exists, load_from_state_file, make_zip):
         """
         Tests connection.
 
@@ -410,6 +413,7 @@ class TestConnClass:
             log_location: location of the logs
             log_file_exists: boolean indicating if log file exist
             load_from_state_file: Defines if the data should be loaded from saved files
+            make_zip: boolean indicating if log file should be zipped
 
         """
         # Arrange
@@ -425,7 +429,7 @@ class TestConnClass:
         if not valid_json:
             self.service.instance.network.uuid = None
 
-        set_up_log(self, log_file_exists, 1)
+        set_up_log(self, log_file_exists, make_zip, 1)
 
         def send_log():
             self.service.event_storage.send_log(self.service.socket)
@@ -1053,9 +1057,10 @@ class TestSendThreadClass:
     @pytest.mark.parametrize("file_size", [2, 1, 0])
     @pytest.mark.parametrize("limit_action", [event_storage.REMOVE_OLD])
     @pytest.mark.parametrize("log_file_exists", [True, False])
+    @pytest.mark.parametrize("make_zip", [True, False])
     def test_send_thread_success(self, messages_in_queue, value, log_offline,
                                  connected, log_location, file_size, limit_action,
-                                 log_file_exists):
+                                 log_file_exists, make_zip):
         """
         Tests sending message.
 
@@ -1089,7 +1094,7 @@ class TestSendThreadClass:
             self.service.socket.sending_queue.put(reply)
         self.service.socket.my_socket.send = Mock(side_effect=KeyboardInterrupt)
         self.service.socket.connected = connected
-        file_path = set_up_log(self, log_file_exists, file_size)
+        file_path = set_up_log(self, log_file_exists, make_zip, file_size)
 
         # Act
         try:
@@ -1130,9 +1135,10 @@ class TestSendThreadClass:
     @pytest.mark.parametrize("file_size", [2, 1, 0])
     @pytest.mark.parametrize("limit_action", [event_storage.REMOVE_OLD])
     @pytest.mark.parametrize("log_file_exists", [True, False])
+    @pytest.mark.parametrize("make_zip", [True, False])
     def test_send_thread_report(self, messages_in_queue, value, log_offline,
                                 connected, log_location, file_size, limit_action,
-                                log_file_exists):
+                                log_file_exists, make_zip):
         """
         Tests sending message.
 
@@ -1167,7 +1173,7 @@ class TestSendThreadClass:
             self.service.socket.sending_queue.put(reply)
         self.service.socket.my_socket.send = Mock(side_effect=KeyboardInterrupt)
         self.service.socket.connected = connected
-        file_path = set_up_log(self, log_file_exists, file_size)
+        file_path = set_up_log(self, log_file_exists, make_zip, file_size)
 
         # Act
         try:
@@ -1209,9 +1215,10 @@ class TestSendThreadClass:
     @pytest.mark.parametrize("file_size", [2, 1, 0])
     @pytest.mark.parametrize("limit_action", [event_storage.REMOVE_OLD])
     @pytest.mark.parametrize("log_file_exists", [True, False])
+    @pytest.mark.parametrize("make_zip", [True, False])
     def test_send_thread_failed(self, messages_in_queue, value, log_offline,
                                 connected, log_location, file_size, limit_action,
-                                log_file_exists):
+                                log_file_exists, make_zip):
         """
         Tests sending message.
 
@@ -1245,7 +1252,7 @@ class TestSendThreadClass:
             self.service.socket.sending_queue.put(reply)
         self.service.socket.my_socket.send = Mock(side_effect=KeyboardInterrupt)
         self.service.socket.connected = connected
-        file_path = set_up_log(self, log_file_exists, file_size)
+        file_path = set_up_log(self, log_file_exists, make_zip, file_size)
 
         # Act
         try:
@@ -1287,9 +1294,10 @@ class TestSendThreadClass:
     @pytest.mark.parametrize("limit_action", [event_storage.REMOVE_OLD])
     @pytest.mark.parametrize("log_file_exists", [True, False])
     @pytest.mark.parametrize("upgradable", [True, False])
+    @pytest.mark.parametrize("make_zip", [True, False])
     def test_send_thread_reconnect(self, messages_in_queue, valid_message, log_offline,
                                    connected, log_location, file_size, limit_action,
-                                   log_file_exists, upgradable):
+                                   log_file_exists, upgradable, make_zip):
         """
         Tests sending message.
 
@@ -1328,7 +1336,7 @@ class TestSendThreadClass:
             self.service.socket.sending_queue.put(reply)
         self.service.socket.my_socket.send = Mock(side_effect=KeyboardInterrupt)
         self.service.socket.connected = connected
-        file_path = set_up_log(self, log_file_exists, file_size)
+        file_path = set_up_log(self, log_file_exists, make_zip, file_size)
 
         # Act
         try:
@@ -1371,9 +1379,10 @@ class TestSendThreadClass:
     @pytest.mark.parametrize("file_size", [2, 1, 0])
     @pytest.mark.parametrize("limit_action", [event_storage.REMOVE_OLD])
     @pytest.mark.parametrize("log_file_exists", [True, False])
+    @pytest.mark.parametrize("make_zip", [True, False])
     def test_send_thread_control(self, messages_in_queue, valid_message, log_offline,
                                  connected, log_location, file_size, limit_action,
-                                 log_file_exists):
+                                 log_file_exists, make_zip):
         """
         Tests sending message.
 
@@ -1413,7 +1422,7 @@ class TestSendThreadClass:
             self.service.socket.sending_queue.put(reply)
         self.service.socket.my_socket.send = Mock(side_effect=KeyboardInterrupt)
         self.service.socket.connected = connected
-        file_path = set_up_log(self, log_file_exists, file_size)
+        file_path = set_up_log(self, log_file_exists, make_zip, file_size)
 
         # Act
         try:
@@ -1455,9 +1464,10 @@ class TestSendThreadClass:
     @pytest.mark.parametrize("file_size", [2, 1, 0])
     @pytest.mark.parametrize("limit_action", [event_storage.REMOVE_OLD])
     @pytest.mark.parametrize("log_file_exists", [True, False])
+    @pytest.mark.parametrize("make_zip", [True, False])
     def test_send_thread_delete(self, object_name, messages_in_queue, log_offline,
                                 connected, log_location, file_size, limit_action,
-                                log_file_exists):
+                                log_file_exists, make_zip):
         """
         Tests sending DELETE message.
 
@@ -1518,7 +1528,7 @@ class TestSendThreadClass:
         self.service.socket.my_socket.send = Mock(side_effect=KeyboardInterrupt)
         self.service.socket.add_id_to_confirm_list = Mock()
         self.service.socket.connected = connected
-        file_path = set_up_log(self, log_file_exists, file_size)
+        file_path = set_up_log(self, log_file_exists, make_zip, file_size)
 
         # Act
         try:
